requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Templating
jinja2>=3.1.3

# Configuration
python-dotenv>=1.0.0

# Google Gemini AI (new SDK)
google-genai>=1.0.0

# HTTP client for GitHub API
httpx>=0.27.0
